backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-openai==0.2.14
langgraph==0.2.38
langgraph-checkpoint-sqlite==0.1.3
python-dotenv==1.0.1
pydantic==2.9.2
langchain-anthropic==0.3.0
